+++ R/zzz.R
@@ -1,3 +1,3 @@
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("GARCOM v0.0.0.9005 package has been loaded")
+  packageStartupMessage("GARCOM v1.0.0 package has been loaded")
 }