+++ NEWS.md
@@ -1,14 +1,14 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 
-GARCOM 0.0.0.9003 Date Aug 18 2020
+`GARCOM 0.0.0.9003 Date Aug 18 2020`
 - Added impute feature to the functions for missingness. We allow mean and median 
 
-GARCOM 0.0.0.9002 Date Aug 12 2020
+`GARCOM 0.0.0.9002 Date Aug 12 2020`
 
 - New features to subset individuals and SNPs
 
-GARCOM 0.0.0.9001 Date Aug 07 2020
+`GARCOM 0.0.0.9001 Date Aug 07 2020`
 - CRAN checks and changes in code for Ubuntu Linux 16.04 LTS, R-release, GCC
 - Column assign step was taken care for IIDs in gene_annot_counts function