+++ NEWS.md
@@ -1,6 +1,6 @@
 * `NEWS.md` file tracks changes made to the package.
 
-`GARCOM 1.2.0 Date Sept 09 2020`
+`GARCOM 1.2.0 Date Sept 10 2020`
 - Minor addition of features with VCF counts: filter genes
 
 `GARCOM 1.1.0 Date Sept 09 2020`